cleanup printf format warnings.

+++ gno/stat.c
@@ -208,7 +208,7 @@ int KERNfstat(int *ERRNO, struct stat *s_buf, int fd) {
 
     disableps();
     if (kp->gsosDebug & 16)
-        fprintf(stderr, "fstat: fd: %d s_buf: %06lX\n", fd, s_buf);
+        fprintf(stderr, "fstat: fd: %d s_buf: %06lX\n", fd, (unsigned long)s_buf);
     if ((fd == 0) || ((fp = getFDptr(fd)) == NULL) || (fp->refNum == 0)) {
         *ERRNO = EBADF;
         enableps();
@@ -256,7 +256,7 @@ int KERNlstat(int *ERRNO, struct stat *s_buf, const char *filename) {
     int rc;
 
     if (kp->gsosDebug & 16)
-        fprintf(stderr, "stat: lpath: %s s_buf: %06lX\n", filename, s_buf);
+        fprintf(stderr, "stat: lpath: %s s_buf: %06lX\n", filename, (unsigned long)s_buf);
     rc = statCommon(filename, s_buf);
     if (rc)
         *ERRNO = rc;
@@ -267,7 +267,7 @@ int KERNstat(int *ERRNO, struct stat *s_buf, const char *filename) {
     int rc;
 
     if (kp->gsosDebug & 16)
-        fprintf(stderr, "stat: path: %s s_buf: %06lX\n", filename, s_buf);
+        fprintf(stderr, "stat: path: %s s_buf: %06lX\n", filename, (unsigned long)s_buf);
     rc = statCommon(filename, s_buf);
     if (rc)
         *ERRNO = rc;
